% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_peer.R
\name{prep_peer}
\alias{prep_peer}
\title{prep_peer}
\usage{
prep_peer(
  ps_export,
  gradebook_export,
  roster_from_grouper,
  activity_name,
  total_points = 1
)
}
\arguments{
\item{ps_export}{The csv exported from peerScholar.}

\item{gradebook_export}{The csv exported from your Canvas gradebook. Must have the assessment you want to process as a column.}

\item{roster_from_grouper}{The class roster, exported from the UT Advanced Group tool. Could be any csv with the columns \code{Student Number} and \code{Email} to allow you to match as peerScholar uses email as the only uniique ID and the Canvas gradebook needs student ID.}

\item{activity_name}{The name of the activity you are grading, exactly as shown in the gradebook csv.}

\item{total_points}{The number of points the assessment is out of on Canvas. Defaults to 1 so the grade is just a proportion.}
}
\value{
Writes a csv of the form "import_acitivity name_date.csv" into the current directory.
}
\description{
prep_peer
}
